drift-backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML
openai==1.12.0
google-generativeai==0.3.2
langchain==0.1.6
langchain-openai==0.0.5
langchain-google-genai==0.0.6

# Auth
pyjwt==2.8.0
httpx==0.26.0

# Database
supabase==2.3.4

# Utils
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
